DEVICE-54 - Support bundling with esbuild
Previously, this module was using dynamic module imports for vendor files.
Systems that use esbuild for bundling were unable to use this module as
esbuild was unable to resolve which files to include in its bundle. This
project has been updated to generate a `src/vendors.js` file based on the
contents of the `vendors/` directory. `src/vendors.js` is imported into other
project utilities, which will force the contents to be included in bundles.